Updates to applyWavelengthSolution and barycentricCorrect
The structure of applyWavelengthSolution has had to be altered, due to
the current limitations of DRAGONS. The ``gemini_tools.make_lists``
functions always expects a single calibrator per ``adinput`` file, and
so crashes out for the wavelength solution, which may have up to two
arcs per ``adinput``. Therefore, the calibrators are now fetched on a
per-``adinput`` basis. This is less efficient than doing a bulk
calibrator query, but currently necessary.

The output extension for the wavelength solution has been renamed to
WAVL, to prevent confusion about the meaning of WFIT.