Change _regrid_spect method back to scipy.interpolate
The numpy.interp function shows some weird behaviour; the scipy
implementation appears to be a lot more stable.

Note that close inspection of the re-gridded sensitivity curve shows
that, on a fine scale, the interpolation isn't great; this is probably
a result of scipy's inability to handle the NaN's which appear in the
data should any elements of the flat come back as zeros. Need to work
out how to address this.