Update to arc calibrator retrieval
The method for retrieving arc calibrators for the calculation of a
wavelength solution has been altered to guarantee one or less arcs
taken before the science observation, and one or less arcs taken
afterwards.

It was necessary to implement a somewhat-complex workaround, given the
calibrations system is designed to return calibrators based on simple
time requirements (e.g. range, before or after), but has no way to
get the calibrators which best bracket a science frame.

The alterations necessary are as follows:
- There is now an extra AstroData descriptor, ``want_before_arc``, which
  will always be ``None`` *except* during arc retrieval, when it will
  be set to True or False to denote whether we want the calibration
  system to return the 'before' or 'after' arc;
- A helper function, _request_bracket_arc, has been added to the
  GHOST primitives to streamline the process of turning this flag on and
  off, and requesting the arcs;
- The __init__ function of the CalibrationGHOST class in the
  calibrations system has been expanded to suck in the
  ``want_before_arc`` information, and the arc matching function has
  been updated to look for that info.

Looking for the arc calibrator(s) for an AstroData object without the
``want_before_arc`` descriptor (i.e. without the special header keyword
``ARCBEFOR``) will still function as before.