Alterations to allow 'full reduction' tests run in the same tmpdir
The method for running 'full reduction' tests will be to sequentially
execute a series of small recipes, checking the outputs at each step
(or, alternatively, using the writeOutputs primitive frequently to
capture certain steps in the reduction process). This technique allows
us to:

- Limit ourselves to only doing one full reduction of the data per test
  run;
- Delete extraneous data files after each test, limiting the pressure
  on the test user's hard drive space;
- Allow the use of a single calibrations setup per test run

Updates also include marking the full reduction tests so they can
be run/skipped easily as a group, and adding teardown code to the
test fixtures.