All arc & standard slit calibrators now processed during testing
This was difficult to do in a fixture sense, because the present
working directory doesn't change to the temporary directory until
we are already inside the fixture. Therefore, the fixture is
parametrized in terms of slit type and resolution, and then each arc
and standard matching those parameters is processed and tested
individually using for loops.