WIP - test to determine if absent PIXELMODEL error being caught correctly.
The current issue is that the RecipeSystem appears to be capturing
the error itself, writing it to the reduction log, and killing the
reduction gracefully. Have contacted KL to see if this behaviour can
be modified or turned off. Otherwise, may need to parse the reduction
log to ensure that the correct errors have been raised.