WIP - tests for PIXELMODEL absence in extractProfile
Have had to switch over to directly using the PrimitiveMapper and
RecipeMapper to run the data reduction, to allow the errors
raised by the primitive to be raised and captured by pytest.

There is currently an issue around the accessing of calibrators,
despite them being provided as per the mapper documentation. Needs
further investigation.