Skip PIXELMODEL related test for now
The test now runs outside the main Reduce() class by utilizing
PrimitiveMapper and RecipeMapper. However, despite calibrators
being specifically defined, the Calibration System (which doesn't
mesh with the API yet) is still being called, which generates an
IOError (calibration not found) and kills the test. Need to figure
out how to bypass this to get to the errors we actually want,
or just switch over to parsing the output log to run the test.