Alter the behaviour of ORIGFILE PHU keyword in un-bundling
Updated _write_newfile to reset the ORIGFILE header keyword of
un-bundled files to be the filename of the new, un-bundled file.
Otherwise, the original name of the MEF bundle is used, which
causes issues when a update_filename(strip=True) command is issued.
update_filename now seems to base the new filename on the
ORIGFILE keyword, which will cause all sorts of conflicting filename
problems if that keyword still holds the MEF bundle filename.